reported bugs fixed

+++ app/src/main/java/com/breadwallet/presenter/entities/PaymentRequestEntity.java
@@ -28,10 +28,10 @@ public class PaymentRequestEntity {
     public static final String TAG = PaymentRequestEntity.class.getName();
 
     public final String[] addresses;
-    public final double amount;
+    public final long amount;
     public final String cn;
 
-    public PaymentRequestEntity(String theAddress[], double theAmount, String theCn) {
+    public PaymentRequestEntity(String theAddress[], long theAmount, String theCn) {
         addresses = theAddress;
         amount = theAmount;
         cn = theCn;
